Multiply devices support: - devicesService moved to services
