+++ src/pages/collection/tracks.jsx
@@ -3,7 +3,7 @@ import { useEffect, useState, useRef, useCallback } from "react";
 import LongPressLink from "../../components/LongPressLink";
 import Image from "next/image";
 import SuccessAlert from "../../components/SuccessAlert";
-import { getCurrentDevice } from "@/lib/device";
+import { getCurrentDevice } from "@/services/deviceService";
 export const runtime = "experimental-edge";
 
 const LikedSongsPage = ({
