+++ src/services/index.js
@@ -2,3 +2,4 @@ export * from './albumService';
 export * from './artistService';
 export * from './playlistService';
 export * from './radioService';
+export * from './deviceService';
